Refactored ElbowMove to support isTallPiece lambda
Defining pieces which cannot be jumped over by the current
expectations of what a `Piece` is leaves room for bloat later
on. If we wanted to create a game mode where pieces grew taller
as they collected kills, this would require changes to the
ElbowMove class. This creates messy move logic.

Using a lambda to express whether or not a piece is tall is
far more expressive.